requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
tensorflow==2.16.1
matplotlib==3.8.4
pillow==10.3.0
python-multipart==0.0.9
typing-extensions>=4.10.0
